a360-data-science/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
attrs==23.2.0
aws-cdk-asset-awscli-v1==2.2.220
aws-cdk-asset-kubectl-v20==2.1.3
aws-cdk-asset-node-proxy-agent-v6==2.1.0
aws-cdk-aws-lambda-python-alpha==2.176.0a0
aws-cdk-cloud-assembly-schema==39.2.2
aws-cdk-lib==2.176.0
boto3==1.36.3
botocore==1.36.3
cattrs==24.1.2
cdk-nag==2.34.23
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
constructs==10.4.2
dill==0.3.9
docker==7.1.0
fastapi==0.115.6
google-pasta==0.2.0
h11==0.14.0
idna==3.10
importlib-metadata==6.11.0
importlib-resources==6.5.2
jiwer==4.0.0
jmespath==1.0.1
jsii==1.106.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.17
numpy==1.26.4
omegaconf==2.2.3
packaging==24.2
pandas==2.2.3
pathos==0.3.3
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
protobuf==5.29.3
psutil==6.1.1
publication==0.0.3
pyarrow==20.0.0
pydantic==2.10.5
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2
rapidfuzz==3.13.0
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.11.1
sagemaker==2.237.3
sagemaker-core==1.0.19
schema==0.7.7
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
starlette==0.41.3
tblib==3.0.0
tqdm==4.67.1
typeguard==2.13.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
